+++ tcob/src/gfx/ui/Scrollbar.cpp
@@ -174,12 +174,12 @@ void scrollbar::calculate_value(point_f mp)
     case orientation::Horizontal: {
         f32 const tw {_barRectCache.Thumb.width()};
         f32 const width {rect.width() - tw};
-        if (width > 0) { frac = (mp.X - _dragOffset.X - (tw / 2)) / width; }
+        if (width > 0) { frac = (mp.X - static_cast<f32>(_dragOffset.X) - (tw / 2)) / width; }
     } break;
     case orientation::Vertical: {
         f32 const th {_barRectCache.Thumb.height()};
         f32 const height {rect.height() - th};
-        if (height > 0) { frac = (mp.Y - _dragOffset.Y - (th / 2)) / height; }
+        if (height > 0) { frac = (mp.Y - static_cast<f32>(_dragOffset.Y) - (th / 2)) / height; }
     } break;
     }
